Moved `Canvas` and drawing logic into new `Screen` SwiftUI component.
