+++ happiNESsApp/ContentView.swift
@@ -23,30 +23,8 @@ struct ContentView: View {
     //     https://www.hackingwithswift.com/quick-start/swiftui/how-to-detect-and-respond-to-key-press-events
     @FocusState private var focused: Bool
 
-    static let width: Int = 32
-    static let height: Int = 32
-    static let scale: Double = 10.0
-
     var body: some View {
-        Canvas {graphicsContext, size in
-            for y in 0 ..< Self.height {
-                for x in 0 ..< Self.width {
-                    let pixel = CGRect(
-                        x: Double(x)*Self.scale,
-                        y: Double(y)*Self.scale,
-                        width: Self.scale,
-                        height: Self.scale)
-
-                    let nesColor = console.screenBuffer[y*Self.width + x]
-                    let color = Color(nesColor: nesColor)
-
-                    graphicsContext.fill(Path(pixel), with: .color(color))
-                }
-            }
-        }
-        .frame(
-            width: CGFloat(Self.width) * Self.scale,
-            height: CGFloat(Self.height) * Self.scale)
+        Screen(screenBuffer: console.screenBuffer)
         .padding()
         .focusable()
         .focused($focused)
